몬스터 스폰 시 공중에서 스폰되지 않도록 조절

+++ shared/NetworkShared/Common/MapData.cs
@@ -20,7 +20,7 @@ namespace NetworkShared.Common
 
     public class Layer
     { 
-        public List<byte> Data { get; set; }
+        public List<short> Data { get; set; }
         public int Id { get; set; }
         public string Name { get; set; }
         public int Opacity { get; set; }
